Moved /addresses/names to separate file
